Implement sharded mode execution for RepoMetadataLogger
Summary: The earlier version of `RepoMetadataLogger` tool was written in Python due to which we had to manually separate out the set of repos into X number of chronos jobs. This approach is error prone and not scalable to thousands of Git repos. Since the current version of the job in written in Rust, we can use our Sharding framework to deploy it to Tupperware where it will be automatically sharded

Differential Revision: D56353208

fbshipit-source-id: f065eaba3f85dff7b193e8fe657dbce7f1a3f827