Configuration for sleep time between incremental iterations of RepoMetadataLogger
Summary: We want to be able to run `RepoMetadataLogger` in incremental mode where we can process only subset of the repo instead of processing the whole repo in every iteration. Since the incremental operation will go over a much smaller subset of file and directories, we want to run the job more frequently than a day but at the same time we don't want it running every 10-20 minutes. This diff allows us to configure how long we want to wait before every iteration for a repo.

Differential Revision: D56421891

fbshipit-source-id: 7ba9afec35f49ed876488cf5f3f83527964b6b2e