merge: add max-distance config to limit distant merges
Summary:
Mononoke bonsai changesets can grow large when two commits that have a distant common ancestor are merged.  The bonsai changeset ends up containing the union of all changes on each side of the merge since the common ancestors.

To mitigate this, add a configuration option that will let us prevent such distant merges.  We define "merge distance" as the number of commits between the common ancestors of the merge parents (if any) and the merge parents themselves.  If there is no such common ancestor, then the merge is of disjoint commit graphs, and the distance is zero.

Reviewed By: sggutier

Differential Revision: D56472179

fbshipit-source-id: a7d91ab3b55d412e46479edceeabc40855b15eff