x_repo_bookmarks_validator: fix flakiness when large repo bookmark cache is more stale than small repo
Summary:
The x_repo_bookmarks_validator finds bookmarks that point to non-matching commits in small and large repos, then for each of them it queries the large repo bookmark update log to try to find a log entry that's not "too old" (its timestamp isn't too old compared to the entry corresponding to the large repo commit) and that corresponds to the commit in the small repo. If it can't find it then it reports this as a inconsistency.

The initial query for the bookmarks goes through the bookmarks cache. This sometimes results in seeing a small repo bookmark that is newer than the large repo bookmark, and since D25196760 we exclude entries in the update log that are newer than the large repo commit which means we don't find an entry that corresponds to the small repo commit and we report a false positive.

To fix this let's not exclude newer update log entry while still comparing timestamps to the log entry that corresponds to the large repo commit. This should avoid re-introducing the issue D25196760 was addressing while avoiding reporting false positives for recent small repo bookmark movements.

Reviewed By: markbt

Differential Revision: D56485312

fbshipit-source-id: aa2a6f2755a7217d7a523cd019ac754510cfe31d