Allow websockets to connect to the current browser's port of choice (if available)
Summary:
Adds a check for the current window's protocol. Defaults to ws:// but will switch to wss:// in secure contexts. This assumes the upstream server will handle converting from wss to ws.

Test plan:
Import github project to replit. Run a server by doing:
`sl web --no-app`

open the replit webview with the token, see if it loads the smartlog.