Live Git Sync: Log and skip sync for non-existent repos
Summary:
Before this diff, both `remote_gitimport` and `git_move_bookmark` would fail as the repo is not found.
This would block a push for the former and return a failure to the client for the latter, based on the
Git hooks they are called in.

Instead, let's log the error to Scuba but have the hooks succeed.

This will allow us to enable the Live Sync hooks by default for all repos. As repos get onboarded in the configs,
Live Sync will be able to start immediately.

This fits in a larger picture where I will also add a bypass git config for the hooks so that we can still
decide to opt-out of Live Sync if required.

Reviewed By: YousefSalama

Differential Revision: D56521512

fbshipit-source-id: ce560619b638d7c5f2b5638b5b367558b2a1fd4c