Pass a title if none provided when splitting
Summary:
If you try to split a commit, and delete the title from one of the commits, it can cause issues in `sl`. Let's force a title to be used in this case.

Similarly, if you use a commit message template, but delete the title of a commit, then you can end up with the commit template field becoming the title, which looks wrong. Let's use the fallback title in this case too.

We could consider using the same title format when you make a commit, like "split of ...", but if they deleted that they probably don't want that title.

Reviewed By: sggutier

Differential Revision: D56541639

fbshipit-source-id: 3bc7961d9627816e5e47d217dabb6e7968ddecc6