Handle commits with no title when parsing commits
Summary:
If you have a commit with no title or description, our commit parsing would `.trim()` the commit away entirely, and then there would be not enough fields, so we'd skip the commit as malformed.

Since we're using all trailing lines after the last field before the COMMIT_END_MARK as the description, we should not trim it.

Reviewed By: sggutier

Differential Revision: D56542648

fbshipit-source-id: c684c654d01ddb74f6f13e22b0c573e9628e2dae