git/protocol: make packfile generation generic over GitDeltaManifest version
Summary: We want to create a new version of `GitDeltaManifest` that uses git's diff algorithm to generate deltas, and we need to be able to use it alongside the current version. In order to do this, this diff creates traits that encapsulate what is needed out of `GitDeltaManifest` and makes the packfile generation code generic over them. This diff also implements a `fetch_git_delta_manifest` method to get whatever `GitDeltaManifest` version is configured in derived data types repo config.

Reviewed By: markbt

Differential Revision: D56578016

fbshipit-source-id: a6bf23836ebe81c421e99cb44b0b652aa74c5394