tests: be more forgiving of `sl goto` output on EdenFS .t tests
Summary:
On D56220001 EdenFS will be enabled for one of our `sl goto` tests. On EdenFS we
usually don't output the number of files that have been updated, and this is
expected behavior. This diff will help us to not having to add a special case
for EdenFS on each case.

Reviewed By: quark-zju

Differential Revision: D56382866

fbshipit-source-id: 0f51f7514918b3c983a5622f63f33b0153713947