tests: enable EdenFS for test-rust-checkout.t
Summary:
One of the initial reasons for us to enabling EdenFS on .t tests was to test
differing behavior between EdenFS and non-EdenFS on various commands at the
same time. The `goto` command is known for having differences between those
two, so let's actually test this. This test in particular also has the
advantage of using the "proper" way of initializing and conducting tests
(i.e., using `modernclient` and `newclientrepo`), which also makes it a
good candidate for this.

Enabling EdenFS on this test already surfaced some differences, which will
be fixed later.

Reviewed By: muirdm

Differential Revision: D56220001

fbshipit-source-id: a5d36881c9d15ea5b52c9d2820f01f4796b3b2bc