Use ButtonDropdown selector to quickly change cwd
Summary:
If there's more than one cwd available, make the cwd tooltip into a ButtonDropdown, so you can quickly change cwds. This is a bit of a shortcut to the larger UI. This addresses a complaint I've heard that it's harder to change cwd, which some repos do a lot.

This UX pattern is a bit peculiar I suppose, but it feels intuitive enough to me. Open to any opinions about this.

Reviewed By: quark-zju

Differential Revision: D56498233

fbshipit-source-id: 6b909292a24eab4a6a1408e22c8534f24c743a78