tests: further modernize test-eden-clone.t
Summary:
This is done in preparation for D56220000. In all of our new kind of tests,
`modernclient` and `newclientrepo` should be used instead of the `eagerepo`
from `tinit.sh` since this function causes a number of issues when used.

Additionally, a relatively new feature, updating on clone, is tested here.

Reviewed By: muirdm

Differential Revision: D56220003

fbshipit-source-id: bb54a016b382e2eec138618b7e8c9118a3fb14cf