smartlog: fix smartlog-default-command hint
Summary:
The hint was firing when just running "sl". This was because the Rust dispatcher was setting Python's `sys.argv` to its args post-processing (which includes adding the default command "sl" on to the list of args). This made it impossible for Python to know if the user typed "sl" or "sl sl" (always appearing as the latter).

Fix by initializing the Python interpreter with the original args. We still pass the processed args to `sapling.run()`, so this should only affect the value of `sys.argv`.

Reviewed By: evangrayk

Differential Revision: D56640367

fbshipit-source-id: 43f2e20edccc54750a170319f806ad8dc79d9ca9