Increase drag target on the right of commits
Summary:
If you drag & drop a commit to rebase, we want the drag target to be a bit forgiving. Currently, it's only EXACTLY the title. So if you dragged onto the diff badge or buttons, it wouldn't work.

This often feels incorrect, for example if you have a commit with a long title, you drag from near the end of the title upwards, near remote/main, it won't preview the rebase unless you get exactly onto the bookmark name.

Instead, we can extend the non-layout part of the <DragToRebase> so that it recieves dragEnter events.

This is done the same way we support dragging from the avatar, with a positive padding and negative margin. This keeps layout the same, but increases the area available for mouse events.

Note that we need to be careful doing this to make sure this new area is not stealing mouse events that were intended for buttons or other UI elements. We can do this with z-index, which seems to work pretty well in my testing.

Reviewed By: jakebolam, sggutier

Differential Revision: D56649406

fbshipit-source-id: 7f091ba4cbd3a675597ca2da370dcd3626657d4a