commit cloud: Allow specifying db type
Summary: The commit cloud service needs some extra operations when we execute queries so we're backwards compatible with the existing db. E.g commits are hex encoded in `xdb.commit_cloud`, so when retrieving data we need to be able to decode them to then pass them up the service.

Reviewed By: markbt

Differential Revision: D56193597

fbshipit-source-id: b45e233a09ab581d6aed7a6193e864fae9faf3f1