Auto-run merge drivers before sl continue
Summary:
Add a config flag to control if you want to run `sl resolve --all` before `sl continue`. This runs automated merge drivers to regenerate generated artifacts.

Sometimes `sl continue` would complain, "you have generated artifacts, run `sl resolve --all` to regenerate". We should just do this automatically for you. We provide a setting so you can opt out of the behavior. That might be useful since merge drivers can be expensive sometimes.

I support we should be doing this in sapling itself? Maybe via a flag when called from ISL?

If we do as a separate command, you'll see if it's `sl resolve --all` that's taking a long time when running in the UI (as opposed to `sl continue`). So we could add special UI documentation that shows that explains it's running merge drivers.

For OSS, I think merge drivers may be less common, so I made this false by default. We could reevaluate this later, and anyone who hasn't tweaked the setting would get the new default.

I also made an inline setting checkbox next to the continue/abort buttons to control this. But only for internal users, since it would likely just be clutter for OSS users. If we determine from analytics that everyone always leaves this active, we could just leave this buried in settings and not show it inline.

Reviewed By: quark-zju

Differential Revision: D56642944

fbshipit-source-id: a8ad580eaf98a618360c92b65d46a0101da4ba31