Allow websockets to connect to the current browser's port of choice (if available) (#886)
Summary:
Allow websockets to connect to the current browser's port of choice (if available)
Adds a check for the current window's protocol. Defaults to ws:// but will switch to wss:// in secure contexts. This assumes the upstream server will handle converting from wss to ws.

Pull Request resolved: https://github.com/facebook/sapling/pull/886

Test Plan:
Import github project to replit. Run a server by doing:
`sl web --no-app`

open the replit webview with the token, see if it loads the smartlog.

Reviewed By: quark-zju

Differential Revision: D56526564

Pulled By: evangrayk

fbshipit-source-id: d241e2f2bfc7fb3fb30d9be9e8206970b8d57cd3