Make continue/abort conflict buttons more prominent
Summary:
The continue and abort buttons were "icon" style buttons, unless the continue button was enabled, then it was primary.

But with the new "Configure External Merge Tool" and "Auto run merge drivers" buttons, it makes "Abort" harder to differentiate quickly.

I think making the two primary actions not use "icon" style makes them easier to see among the other buttons now.

We may revisit this again in the future, but for now I think this is an improvement.

Reviewed By: quark-zju

Differential Revision: D56658486

fbshipit-source-id: 3e23bb901af4582f901c6ad12ef65775a26038e3