Log performance stats during submodule expansion
Summary:
It will be very useful to know how much time we're spending on each part of the submodule expansion and validation flow.

It will help determine if and where we need to work on improving performance, because the inital import can run for a long time, but the same code will be used by live forward and backsyncer, so it needs to be reasonably fast.

Reviewed By: andreacampi

Differential Revision: D56561379

fbshipit-source-id: 26544914299e2ef022d472a0269920fe26f5b1f4