eden rm: When failing to delete a file, consider ACCESS_DENIED as a maybe fixable problem
Summary:
The original code considers only "fixable" the error code "used by another process", but not "access denied". However, this is the error code we get when the file we want to delete is a binary that is running.

This diff accepts that error code as fixable and therefore will trigger the kill process branch.

This is a subset of D56490217 which I've been asked to split in parts.

Reviewed By: MichaelCuevas

Differential Revision: D56589592

fbshipit-source-id: 6474ea50023131297ade949a8b166cad5cade60e