Distinguish generated file changes from expansion
Summary:
There are scenarios where the expansion of submodules might generate multiple file changes for the same path.

These can fall into two cases:
1. One is from the original bonsai (i.e. from small repo) and another was generated
2. Both of them were generated: one deletion and one change.

We want to make sure that we keep track of which changes were from the original bonsai and which were generated, so we can set the correct order of precedence when creating the final `SortedVectorMap`.

File changes from the original bonsai **always take precedence**. If both of the file changes were generated, the regular change should take precedence over the deletion.

## This diff
Introduces an enum that we use to make sure that original and generated file changes are kept separate and are merged together with the correct order of precedence.

Reviewed By: mitrandir77

Differential Revision: D56693920

fbshipit-source-id: c0de7ae24d67e944350f51c64af1c6bffe207795