Unit test for deleting submodule but keeping its contents
Summary:
One scenario we found during the first test merge using submodule expansion is that a user removes a submodule, but keeps a static copy of it in the same path.

This is currently failing validation, because it expects the entire expansion to be removed if the submodule metadata file is removed.

I synced with mitrandir77 and markbt about this and we got to the conclusion that this validation logic **is incorrect**.
The change made by the user is valid, so we should treat the removal of the submodule metadata file as a valid change. Everything left in the submodule path will be treated as regular file changes.

One issue I expected to see from doing this would be later changes being made to this static copy failing validation because the commit was modifying a path in the submodule dependency map (i.e. theoretically a submodule expansion) without modifying the metadata file.

So I added this scenario in these tests, because they are addressed in the fix on the next diff (D56690832).

Reviewed By: mitrandir77

Differential Revision: D56633699

fbshipit-source-id: 55c558fae26dd0ef9671b4016d43b48a3c929eba