Implement Git host selection for Mononoke GRit
Summary: The Git repos on Mononoke GRit server are deep sharded, i.e. only a subset of repos are present on any one server. This is why we need a routing layer that identifies the correct host for an incoming request for a Git repo. This crate is meant to serve that purpose. It will be used in Metagit Location Service and Metagit Follower during replication

Reviewed By: mitrandir77

Differential Revision: D56671042

fbshipit-source-id: c75c360e41a7d461fb0cdb63ba662d9133714b69