eden: Introduce better parser for output of handle.exe
Summary:
The original code had a very naive parser for the output of handle.exe (basically a whitespace split and an abort if a line with 4 works was found).

This diff introduces a regexp parser that will skip unknown lines.

Note: It introduces the new parser but it doesn't yet use it (left for the next diff in the stack).

This is part of a split of D56490217

Differential Revision: D56593479

fbshipit-source-id: 77c2262f18afc791efb5fdc6b4403a24fd03d33c