commit cloud: Initial implementation for get_references endpoint
Summary:
This is a basic impl, not ready for testing cause we don't correctly have the heads dates since we need to retrieve them from SCS.

This diff only contains the commit cloud service part, edenapi declaration and implementation will come in further diffs.

This diff also adds some data structures like ReferencesData that will be declared in edenapi later and used here. This is just me tryin to keep diffs as "small" as possible.

Reviewed By: mitrandir77

Differential Revision: D56647152

fbshipit-source-id: 0d294cbf5f470f1c2406e8f9fda383fd327a7342