commit cloud: Use changeset handlers in data structures
Summary:
So for commit cloud there's two DBs, `xdb.commit_cloud`(prod) and a sqlite one we use for test and that we're aiming to move to as we move more things over to mononoke.  I'm trying to make these queries as deduped as possible so there's a seamless db usage instead of having to define everything twice.

One of the main differences is that we store hex-encoded commits in `xdb.commit_cloud`, so, the previous diff added functions that encode changesets depending on the db type, and here we just use them in the existing data structures. Technically we could still simply use `HgChangesetId`for sqlite and bytes for mysql. But again, I'm trying to make this as seamless and possible and abstract all of the minor type casting within the db and avoid duplication.

Differential Revision: D56597877

fbshipit-source-id: a02ea5ee0f55e06baaae9b45532e2da2251a565b