Add Telemetry to findGlob
Summary:
# Context
Suffix queries are a type of query that tend to take a very long time. We would like to determine how long this type of glob queries takes in order to determine impact of improvements to this type of query.

# This Diff
This diff adds telemetry to findGlobs that adds additional logging if a suffix glob query is present.

# Technical Details
This telemetry is done via creating a SuffixLogRequestScope that creates the log when it gets cleaned up.
Logging is done to XLOG via a string output, and Scuba via the SuffixGlob object

Reviewed By: jdelliot

Differential Revision: D55259063

fbshipit-source-id: 1d802832395089b35139cc2c00fafd1b2420fc26