manifest: use the same test manifest for bonsai and implicit deletes
Summary:
These tests use a different manifest type than the other tests, as well as a custom store.  This is error-prone, for example `test_unfold_dir_from_dirs` has a cyclic manifest at `2`, which should be impossible.  It's also an unnecessary duplication of code.

Switch it to use the same test manifest as the rest of the tests.

Differential Revision: D56569171

fbshipit-source-id: ea69f3016ed0272540dd04f9317bd3750a18ccbb