Log memory stats inside an expensive method
Summary: I am initially adding this instrumentation to just one method to get some experience in practical use. I chose this method because it's pretty resource-intensive. The qps is low enough that I'm not worried about the overhead, and the JK allows us to turn this off if it gets in the way at all. It might be worth memoizing `memory::get_stats` with a short TTL (well under a second though, or it wouldn't be useful with sudden spikes).

Reviewed By: clara-9

Differential Revision: D56701239

fbshipit-source-id: 3f2e525df983e5e8414da45c6de885c268684b7f