eden rm: Stop adb server / buck2
Summary:
When eden rm fails due to resources being locked a usual reason is adb server.

We can't kill adb because it runs as admin, but luckily we can ask it to stop.

Since it's started automatically when needed, this is harmless.

If it wasn't running, it's just no op.

Same with buck2.

Also - fix a missing arg in an instantiation of WinFileHandlerReleaser :-(

This diff is part of a split of D56490217

Reviewed By: MichaelCuevas

Differential Revision: D56590257

fbshipit-source-id: 156d3ed72039cc8783a719640551f114abc39671