Enforce memory limits
Summary:
Introducing a whole new exception so we can configure different behavior in ServiceRouter, as well as keeping separate metrics.

This is a crude mechanism for now, but it should be enough to distribute the load better when getting overloaded by a few calls. A followup diff will make the behavior configurable by method and repository.

Reviewed By: clara-9

Differential Revision: D56715749

fbshipit-source-id: 775f52d716c5984132a02c799f1e501440887e85