Automatically derive fsnodes during initial import
Summary:
We need to derive fsnodes for the commits as they're imported, because otherwise we'll be rederiving it for all ancestors of the commits being processed every time, leading to O(n2) complexity.

This is because without this change we request derivation using the large repo `InMemoryRepo`, which doesn't persist the data to the production blobstore afterwards.

I created T187426389 to go further and automatically derive everything for the commits as they go.

Differential Revision: D56706360

fbshipit-source-id: d107cb32673a8259e445d4c71b2d24715b6944a8