Return early on validation
Summary:
I noticed that import was going very slow, but that was being spent on validation of submodules that weren't even being modified by a changeset...
So I looked into it and it turns out we were always deriving the fsnodes for the large bonsai, even when we didn't need to.

This doesn't really solve the problem of slow derivation, but saves some unnecessary delays for most changesets.

Differential Revision: D56701028

fbshipit-source-id: 158c3992b2102a4642930329064732b48d4b32a5