fix readdir tests on FilteredFS
Summary:
# Context

FilteredFS mounts use FilteredBackingStores and therefore they use FilteredObjectIDs instead of underlying ObjectIDs.  That means the getAttributesFromFiles endpoint will spit out FilteredObjectIDs instead of ObjectIDs. This test previously checked only for ObjectIDs. For FilteredFS mixins of this test we should expect to see FOIDs instead of OIDs.

# This diff

Adds a utility to translate underlying ObjectIDs into FilteredObjectIDs based on the type of object that's passed in. This causes all the re-enabled tests to pass.

Reviewed By: jdelliot

Differential Revision: D56593399

fbshipit-source-id: a428782c05617d8ad34f87d8c29796ed36bc5946