FilteredFS: add test case that showcases bug
Summary:
# Context

An XFN partner reported a bug that occurs when a user modifies the filter file that's currently active. In other words, commiting changes to the active filter file leads to unexpected behavior.

**Expectation**

Committing changes to a filter file will cause those changes to be immediately reflected in the repo working copy.

**Actual behavior**

Committing changes to a filter file doesn't actually update the working copy accordingly. Instead, the changes (especially those that cause a directory to go from filtered -> unfiltered) will cause the newly unfiltered files to show up as "missing" in `sl status`.

# Fix

We need to fix the `sl commit` hook to update the working copy when filter changes are commited.

# This diff

Adds a test that showcases the broken behavior. This test will be used to verify the bug fix works as expected.

Reviewed By: sggutier

Differential Revision: D56445402

fbshipit-source-id: 5fd0fb66a3721d0a0696ba741a8a23f737cebde4