FilteredFS: fix commiting filter changes bug
Summary:
# Context

An XFN partner reported a bug that occurs when a user modifies the filter file that's currently active. In other words, commiting changes to the active filter file leads to unexpected behavior.

**Expectation**

Committing changes to a filter file will cause those changes to be immediately reflected in the repo working copy.

**Actual behavior**

Committing changes to a filter file doesn't actually update the working copy accordingly. Instead, the changes (especially those that cause a directory to go from filtered -> unfiltered) will cause the newly unfiltered files to show up as "missing" in `sl status`.

# Fix

We need to fix the `sl commit` hook to update the working copy when filter changes are commited.

# This diff

Fixes the behavior of commiting a filter change so that the filter change is immediately reflected after the commit is made.

Reviewed By: quark-zju

Differential Revision: D56445400

fbshipit-source-id: a79296cd285ea21b577f608361378cbf8a36ad2c