fix doctor tests on FilteredFS mounts
Summary:
# Context

The scm_type of FilteredFS mounts isn't "hg", instead it's "filteredhg" to indicate a FilteredBackingStore is being used instead of just the underlying HgQueuedBackingStore.

This caused Doctor to skip some health checks because the hg-specific logic was gated on `CheckoutConfig.scm_type == "hg"`

# This diff

This diff adds "filteredhg" to the list of scm_type that will trigger hg-specific doctor checks.

Reviewed By: genevievehelsel

Differential Revision: D56593396

fbshipit-source-id: c12b7d9a605a268d95dfeca5bdbff50e47d4cd25