make `eden clone` respect backing store types
Summary:
# Context

If the user supplies a backing-repo path that points to an existing Eden repo, then clone will try to copy that existing repo's CheckoutConfig. This works in most cases, but it will override user supplied options in other cases. One such case was when the user supplied a --backing-repo option. This option was entirely ignored in favor of whatever value was contained in the existing repo's checkout config.

# Solution

We should teach clone to respect the user-supplied option or default to "hg" if no option is supplied.

# This diff

Adds a check to ensure the backing store types match. If not, it replace the backing store field with the option supplied by the user (or "hg" by default)

Reviewed By: genevievehelsel

Differential Revision: D56593400

fbshipit-source-id: 608e26a6556cd6f0df6ed06a963ea1bfbfc0ea77