fix existing getCurrentSnapshotInfo test
Summary:
# Context

If the user supplies a backing-repo path that points to an existing Eden repo, then clone will try to copy that existing repo's config. This works in most cases, but it will override user supplied options in some cases. One such case was when the user supplied a --backing-repo option. This option was entirely ignored in favor of whatever value was contained in the existing repo's checkout config.

# Solution

We should teach clone to respect the user-supplied option or default to "hg" instead.

# This diff

Fixes a test that was originally broken by this bug.

Reviewed By: genevievehelsel

Differential Revision: D56593402

fbshipit-source-id: 2cb47a177669137728172e007980b53a2a03d44e