pymanifest: remove a copy from walkfiles()
Summary:
If we have a pure Rust matcher (which sould always be the case, I think), avoid a copy of the list of files. Since we don't need to give up up the Python GIL, we can create the Python objects as we iterate the manifest results.

This is called from `prefetch()`, so the list of files can be large.

Reviewed By: quark-zju

Differential Revision: D56639144

fbshipit-source-id: 7b477218f3a110776c7dfaf9d2ac621f3dda06a9