derived_data_service: make ZelosDerivationQueue struct correspond to a single repo and a single derived data config
Summary:
The main struct used in the derived data service and the derivation worker for all interactions with the derivation queue is `ZelosDerivationQueue`. This struct consist of a zeus client and a collection of repos (i.e. `MononokeRepos`), which means that there isn't a struct that represent the derivation queue for a single repo. This complicates adding admin commands for inspecting the derivation queue, complicates all interactions with the queue as we have to check that the repo and the derived data config we want to process actually exists in `MononokeRepos` (e.g. https://fburl.com/code/qdja0m09), and overall doesn't fit with the usual Mononoke style of having repos as the central objects and facets for each repo.

This diff tries to fix this by making `ZelosDerivationQueue` correspond to a single repo and a single derived data config, and adding a repo facet `RepoDerivationQueues` which is a map from config name to `ZelosDerivationQueue`. This diff is large because it touches the main object of the derived data service, but it's mostly the result of following and fixing compiler errors.

Reviewed By: singhsrb

Differential Revision: D56766255

fbshipit-source-id: 41bc8fc6961b757b11248789f9d2013a62be0d31