Fix tests run locally with 0 async timeout
Summary:
Somehow setting the asyncUtilTimeout to undefined is being treated as 0ms instead of default = 1000ms. It seems to work properly if the property is missing instead of undefined, so we can use a spead syntax instead.

Not sure why this didn't come up in the original diff, I swear I tested this locally.

Reviewed By: zzl0

Differential Revision: D56836613

fbshipit-source-id: 6e108f603be119be65497e6ee7f29f111069bb4e