Show deleted file conflicts differently in UI and run sl rm to resolve them
Summary:
Now that we have a `conflictType` in the conflict info, we can update the UI to show a "delete" action for deletion conflicts. We can also add a hint about which kind of deletion it is: "deleted in dest" or "deleted in incoming".

The "delete" action needs to both delete the file and resolve it. I prefer using `sl rm` instead of using an fs delete operation, since we'd need to queue it up somehow and handle not allowing deleting files  outside the repo for security, etc.

There is a bug in sl right now that makes our conflict fetching not work immediately after running `sl rm` on the file but before the resolve comes through. This is fixed in D56800824. But since this action immediately resolves the file, this bug isn't a deal-breaker.

The names for the comparison sides are still not ideal, so I suspect this will change in the near future when we revisit conflicts again. But this will do for now I think.

Reviewed By: zzl0

Differential Revision: D56802649

fbshipit-source-id: 7727163414f12830cbd5a9cd52308b1b33dfdc4d