Add conflict type to the conflicts we fetch to know if we need to give deletion actions
Summary:
Conflicts that are deleted in the source/dest are not resolvable in the same way as normal "both changed" conflicts, so we need special detection and handling.

So when fetching conflicts, we compute and store the conflict type as well, based on `exists` in the conflict data.

Then, the UI can use this data to modify which actions it shows. We can also show a hint that the file was deleted

In this diff, I do the server side handling to fetch the conflict type.

In the next diff, I update the UI to use this

Reviewed By: zzl0

Differential Revision: D56802648

fbshipit-source-id: 266aed8805839fb5a7348a6f6c8c84add89378f9