Handle \r in command output
Summary:
Some long lived commands (e.g. those run as merge drivers) can display arbitrary output, including carriage returns for progress bar rendering:
```
Progress 10%\rProgress 11%\rProgress12%\r...
```

These would display somewhat badly in the UI, without collapsing lines, so you'd just see a hundred lines of progress bar output. That's not the end of the world, but it's a bit hard to read.

It's not too hard to handle `\r` in lines when rendering.

Note: there was some previous discussion in D48836829, which seems to have ended in wanting \r to be treated like in this diff, instead of made into newlines.

Reviewed By: zzl0

Differential Revision: D56779045

fbshipit-source-id: fe6f615f79b641964e074815758ab18428a160e5