fix error with resolve internal:dumpjson with a deleted file conflict
Summary:
Bug:
- have a conflict in a deleted file (e.g. rebase A onto B where A deletes file.txt and B modifies file.txt)
- `sl rm file.txt`, in an attempt to use the deleted version of the file (and later mark it as resolved to continue the rebase)
- `sl resolve --tool internal:dumpjson --all` would hit an exception and not print valid json:
```
abort: No such file or directory: /path/to/repo/file.txt
```

This bug is bad becuase it makes us unable to detect merge conflicts at all which leaves ISL in a bad state.

This seems to be from one of the file contexts not being considered absent, and thus we try to read the file contents.

~~The fix here is to wrap the content reading in a `try`.~~

The fix is to check `.lexists` before accessing `.data` when the context is a workingfilectx

Let me know if there's a better way we should be solving this. This is pretty isolated code for ISL's consumption AFAIK.

~~This breaks one tests for a case where a conflicting "file" is a directory in one side of the conflict. The "intended" behavior was to abort there. I don't think this kind of command should abort for expected situations, only for situations where something is very broken. It seems better to return valid json whenever possible. So I just changed the test to be ok with "exists" false, even though it's an existant directory now.~~

Update: The new approach without `try` re-breaks this test, which I guess was intended to be broken before. Still not crazy about having invalid json possible from this command, but I guess it's quite a rare case. We could instead:
- stat the file in addition to `.lexists()`
- go back to using a `try` around the `.data`

Reviewed By: zzl0

Differential Revision: D56800824

fbshipit-source-id: 342927f0d127e857cf0ca698e81d16193f45ba3b