Add local/other hashes to resolve internal:dumpjson
Summary:
It's useful in the ISL UI to know the different sides of a given merge conflict. This is pretty easy to add to the json output we already have.

In the future, we'd like to also be able to find the specific commit that introduced the conflict (not necessarily the dest). That will require more thorough blame and could be expensive, so we'll do that in a separate async call that doesn't block conflicts loading.

However, it's useful to know the hashes immediately so we can show what conflict you're in.

I was thinking of including the "base" of the merge too. It seems this is different per path with conflicts. At least, it's accessed in mergestate for a specific path. I don't know that it's all that useful in ISL so I just skipped it for now.

It also seems for some conflicts these hashes can be null, which is fine.

Reviewed By: quark-zju

Differential Revision: D56850550

fbshipit-source-id: 9fa48963a3dcbbf3008c14767596599b385f1986