Add type for conflict hashes fetched from resolve :dumpjson
Summary:
D56850550 added `"hashes": {"local": "abc", "other": "def"}` hashes to the json output of `sl resolve --tool internal:dumpjson`. This is useful for ISL to show what's being rebased during a conflict.

If you're not yet using a version of sl that has this feature, this would be missing from the JSON, so we make these values nullable. Even if you do have this version of sl, some generated conflicts will not have these values either.

Reviewed By: zzl0

Differential Revision: D56861539

fbshipit-source-id: e35d48a80046fe6d9c7ce7ee675e8dc00ba308a8